SWATCH-1750: Configure docker in docker in Jenkins
