+++ dependencies.gradle
@@ -69,3 +69,4 @@ libraries["webjars-locator"] = "org.webjars:webjars-locator:0.47"
 libraries["wiremock"] = "com.github.tomakehurst:wiremock-standalone:3.0.1"
 libraries["awaitility"] = "org.awaitility:awaitility:4.2.0"
 libraries["splunk-otel-agent"] = 'com.splunk:splunk-otel-javaagent:1.28.0'
+libraries["testcontainers-postgresql"] = 'org.testcontainers:postgresql:1.18.3'
