+++ docker-compose.yml
@@ -2,12 +2,13 @@
 version: '3.1'
 services:
   db:
-    image: docker.io/library/postgres:13
-    command: '-c max_connections=5000'
+    image: quay.io/centos7/postgresql-13-centos7
     environment:
       - POSTGRES_HOST_AUTH_METHOD=trust
+      - POSTGRESQL_MAX_CONNECTIONS=5000
+      - POSTGRESQL_ADMIN_PASSWORD=admin
     volumes:
-      - ./init_dbs.sh:/docker-entrypoint-initdb.d/init_dbs.sh:z
+      - ./init_dbs.sh:/usr/share/container-scripts/postgresql/start/set_passwords.sh:z
     ports:
       - "5432:5432"
   kafka:
