+++ init_dbs.sh
@@ -1,9 +1,13 @@
-#!/bin/sh
-set -e
+#!/bin/bash
+
+_psql () { psql --set ON_ERROR_STOP=1 "$@" ; }
+
+_psql --set=adminpass="$POSTGRESQL_ADMIN_PASSWORD" \
+<<<"ALTER USER \"postgres\" WITH ENCRYPTED PASSWORD :'adminpass';"
 
 for db in rhsm-subscriptions insights; do
-  psql <<-EOSQL
-      CREATE USER "$db";
+  _psql --set=adminpass="$POSTGRESQL_ADMIN_PASSWORD" <<-EOSQL
+      CREATE USER "$db" WITH PASSWORD '$db';
       CREATE DATABASE "$db";
       GRANT ALL PRIVILEGES ON DATABASE "$db" TO "$db";
       ALTER USER "$db" WITH SUPERUSER;
