+++ settings.gradle
@@ -23,6 +23,7 @@ include ':swatch-common-config-workaround'
 include ':swatch-common-resteasy'
 include ':swatch-core'
 include ':swatch-core-test'
+include ':swatch-common-testcontainers'
 include ':swatch-system-conduit'
 include ':swatch-producer-aws'
 include ':swatch-contracts'
