+++ swatch-common-testcontainers/build.gradle
@@ -1,0 +1,7 @@
+plugins {
+    id 'swatch.java-conventions'
+}
+
+dependencies {
+    compileOnly libraries["testcontainers-postgresql"]
+}
