+++ swatch-core-test/build.gradle
